% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_helpers.R
\name{BuildMultigroupDEGlist}
\alias{BuildMultigroupDEGlist}
\title{Build up a DEG data frame for clusterProfiler multi-group ORA analysis}
\usage{
BuildMultigroupDEGlist(Gene, Group, OrgDB = org.Hs.eg.db)
}
\arguments{
\item{Gene}{A vector containing genes.}

\item{Group}{A character of group name}

\item{OrgDB}{Default as org.Hs.eg.db}
}
\value{
A data frame. Following rbind command can help you build up the full DEG list.
}
\description{
Convert symbol, and add up group name in column.
}
\author{
Zhiming Ye
}
