% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuser.R
\name{compareGT}
\alias{compareGT}
\title{2-Group Comparison of enrichment results and further clustering and visualizing}
\usage{
compareGT(
  obj.test,
  obj.ctrl,
  name.test = NULL,
  name.ctrl = NULL,
  ClusterNum = 15,
  P.adj = 0.05,
  force = F,
  nTop = 10,
  ...
)
}
\arguments{
\item{obj.test}{the enriched object from tested group. WARNING: `obj.test` and `obj.ctrl` should come from same database (e.g. GO Biological Process(GOBP)).}

\item{obj.ctrl}{the enriched object from control group. WARNING: `obj.test` and `obj.ctrl` should come from same database (e.g. GO Biological Process(GOBP)).}

\item{name.test}{optional, the name of the testing group. If is `NULL`, the object name of `obj.test` will be used.}

\item{name.ctrl}{optional, the name of the control group. If is `NULL`, the object name of `obj.ctrl` will be used.}

\item{ClusterNum}{how many cluster will be clustered}

\item{P.adj}{p.adjust cut-off. To avoid slow visualization, you can make stricter p-cut off.}

\item{force}{ignore all auto-self-checks, which is useful}

\item{nTop}{keep n top items according to p-adj in each cluster.}

\item{...}{Others options.}
}
\value{
an `EnrichGT_obj` object. slot `enriched_result` contains a data.frame with enriched results. `gt_object` contains `gt` object. you can use `obj@gt_object` to get it and use functions from `gt` like `gtsave`. `gene_modules` is a list containing meta-gene modules of each cluster. `pathway_clusters` contains pathways names in each cluster.
}
\description{
2-Group Comparison of enrichment results and further clustering and visualizing
}
\details{
Execute `obj.test` VS `obj.ctrl` tests, showing pathway overlaps (or differences) and meta-gene modules of test group and control group. Supports ORA and GSEA results (enriched object or data.frame). !WARNING!: `obj.test` and `obj.ctrl` should come from same database (e.g. GO Biological Process(GOBP)).
}
\author{
Zhiming Ye
}
