% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{convert_annotations_genes}
\alias{convert_annotations_genes}
\title{Convert gene annotations from any keys to any keys}
\usage{
convert_annotations_genes(genes, from_what, to_what, orgDB)
}
\arguments{
\item{genes}{gene vector}

\item{from_what}{input type (like "SYMBOL","ENTREZID","ENSEMBL","GENENAME",...), keys should be supported by AnnotationDbi. Search for the help page of AnnotationDbi for further help.}

\item{to_what}{output type (like "SYMBOL","ENTREZID","ENSEMBL","GENENAME",...), keys should be supported by AnnotationDbi. Search for the help page of AnnotationDbi for further help. Can be multiple items E.g. `c("ENTREZID","ENSEMBL","GENENAME")`}

\item{orgDB}{human = org.Hs.eg.db, mouse = org.Mm.eg.db, search BioConductor website for further help}
}
\value{
a data.frame
}
\description{
Convert gene annotations from any keys to any keys
}
