% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_helpers.R
\name{database_from_gmt}
\alias{database_from_gmt}
\title{getGMTFile}
\usage{
database_from_gmt(gmtfile)
}
\arguments{
\item{gmtfile}{gmt file path}
}
\value{
data.frame
}
\description{
read `.gmt` files. You can download them from https://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp
}
\details{
parse gmt form MsigDB file to a data.frame
}
\author{
cited from https://github.com/YuLab-SMU/gson/blob/main/R/GMT.R . The further Cache system is written by Zhiming Ye.
}
