% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_helpers.R
\name{doEnrich}
\alias{doEnrich}
\title{A C++ accelerated universal enrichment analyzer}
\usage{
doEnrich(
  genes,
  database,
  p_adj_methods = "BH",
  p_val_cut_off = 0.5,
  background_genes = NULL,
  min_geneset_size = 10,
  max_geneset_size = 500,
  multi_cores = 0
)
}
\arguments{
\item{genes}{a vector of gene id}

\item{database}{a database}

\item{p_adj_methods}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{p_val_cut_off}{adjusted pvalue cutoff on enrichment tests to report}

\item{background_genes}{background genes. If missing, the all genes listed in the database}

\item{min_geneset_size}{minimal size of genes annotated for testing}

\item{max_geneset_size}{maximal size of genes annotated for testing}

\item{multi_cores}{multi_cores}
}
\description{
A C++ accelerated universal enrichment analyzer
}
