% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_helpers.R
\name{doGSEA}
\alias{doGSEA}
\title{GSEA analysis by EnrichGT}
\usage{
doGSEA(
  genes,
  database,
  min_geneset_size = 10,
  max_geneset_size = 500,
  gseaParam = 1
)
}
\arguments{
\item{genes}{a vector of gene id}

\item{database}{a database}

\item{min_geneset_size}{minimal size of genes annotated for testing}

\item{max_geneset_size}{maximal size of genes annotated for testing}

\item{gseaParam}{pass to fgsea}

\item{p_adj_methods}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{p_val_cut_off}{adjusted pvalue cutoff on enrichment tests to report}
}
\description{
A warpper of fgsea::fgsea
}
