% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuser.R
\name{egt_compare_groups}
\alias{egt_compare_groups}
\title{2-Group Comparison of enrichment results and further clustering and visualizing}
\usage{
egt_compare_groups(
  obj.test,
  obj.ctrl,
  name.test = NULL,
  name.ctrl = NULL,
  ClusterNum = 15,
  P.adj = 0.05,
  force = F,
  nTop = 10,
  method = "ward.D2",
  ...
)
}
\arguments{
\item{obj.test}{the enriched object from tested group. WARNING: `obj.test` and `obj.ctrl` should come from same database (e.g. GO Biological Process(GOBP)).}

\item{obj.ctrl}{the enriched object from control group. WARNING: `obj.test` and `obj.ctrl` should come from same database (e.g. GO Biological Process(GOBP)).}

\item{name.test}{optional, the name of the testing group. If is `NULL`, the object name of `obj.test` will be used.}

\item{name.ctrl}{optional, the name of the control group. If is `NULL`, the object name of `obj.ctrl` will be used.}

\item{ClusterNum}{how many cluster will be clustered}

\item{P.adj}{p.adjust cut-off. To avoid slow visualization, you can make stricter p-cut off.}

\item{force}{ignore all auto-self-checks, which is useful}

\item{nTop}{keep n top items according to p-adj in each cluster.}

\item{method}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{...}{Others options.}
}
\value{
`List` containing multiple `EnrichGT_obj` objects.

The `List` contains objects with overlapped enriched terms, unique enrich terms.
}
\description{
See \code{?egt_enrichment_analysis()}
}
\details{
Execute `obj.test` VS `obj.ctrl` tests, showing pathway overlaps (or differences) and meta-gene modules of test group and control group.

Supports ORA and GSEA results (enriched object or data.frame).

!WARNING!: `obj.test` and `obj.ctrl` should come from same database (e.g. GO Biological Process(GOBP)).
}
\author{
Zhiming Ye
}
