% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicpr.R
\name{egt_comparison_reactor}
\alias{egt_comparison_reactor}
\title{Framework for comparing and analyzing enrichment results
across multiple groups}
\usage{
egt_comparison_reactor(Type = NULL)
}
\description{
Framework for comparing and analyzing enrichment results
across multiple groups
}
\details{
Type should be one of "ORA" or "GSEA".
For functions inside reactor, please see below 'See also' (in the bottom of this doc)
}
\examples{
\dontrun{
# ORA example
reactor <- egt_comparison_reactor("ORA")
reactor$append_enriched_result(ora_result1, "group_1_go")
reactor$append_enriched_result(ora_result3, "group_3_go")
reactor$prefilter_by_p_adj(0.05)
reactor$make_plans(group = c("group_1_go","group_3_go"), use_value = "p")
reactor$find_relationship(Num = 3)
wordcloudFigure <- reactor$fetch_biological_theme()
wordcloudFigure[[1]]
reactor$split_by_cluster()
reactor$do_recluster(ClusterNum = 10)
cls_res <- reactor$get_recluster_result()
relation_df <- reactor$fetch_relationship()

# GSEA can do more with:
gsea_reactor <- egt_comparison_reactor("GSEA")
gsea_reactor$prefilter_by_NES(1.5)
}


}
\seealso{
\code{\link{comparison_reactor_base}} for the base class documentation
\code{\link{comparison_reactor_ora}} for ORA-specific functionality
\code{\link{comparison_reactor_gsea}} for GSEA-specific functionality
}
