% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2.R
\name{egt_fetch_biological_theme}
\alias{egt_fetch_biological_theme}
\title{Generate Biological Theme Wordcloud from Enrichment Results}
\usage{
egt_fetch_biological_theme(x, cluster = NULL, skip_filtering = F, ...)
}
\arguments{
\item{x}{Input object - either a data.frame with enrichment results or an EnrichGT_obj}

\item{cluster}{Cluster name/number (required if x is EnrichGT_obj)}

\item{skip_filtering}{Logical, if TRUE skips filtering of data.frame by p.adjust}

\item{...}{Additional arguments passed to wordcloud_generator2}
}
\value{
ggwordcloud object
}
\description{
Creates a wordcloud visualization from enrichment results, either from a data.frame
or an EnrichGT_obj. For data.frames, filters by p.adjust < 0.05 by default.
}
\author{
Zhiming Ye
}
