% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infAct.R
\name{egt_infer}
\alias{egt_infer}
\title{Infering Pathway or Transcript Factors activity from EnrichGT meta-gene modules}
\usage{
egt_infer(x, DB = "collectri", species = "human")
}
\arguments{
\item{x}{an EnrichGT_obj object.}

\item{DB}{can be "progeny" (the Pathway activity database), or "collectri" (TF activity database)}

\item{species}{can be "human" or "mouse"}
}
\value{
a `compareCluster` result from `clusterProfiler`
}
\description{
Only supports gene symbols. so you must use enrichedObj |> setReadable(OrgDb = xxx,keyType = "ENTREZID") |> EnrichGT() . Do Not supports ENTREZIDs!

[PROGENy](https://saezlab.github.io/progeny/) is a comprehensive resource containing a curated collection of pathways and their target genes, with weights for each interaction.

[CollecTRI](https://github.com/saezlab/CollecTRI) is a comprehensive resource containing a curated collection of TFs and their transcriptional targets compiled from 12 different resources. This collection provides an increased coverage of transcription factors and a superior performance in identifying perturbed TFs compared to our previous.
}
\author{
Zhiming Ye, Saez-Rodriguez Lab (The decoupleR package, https://saezlab.github.io/decoupleR/)
}
