% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objFrame.R
\name{egt_summary}
\alias{egt_summary}
\title{Generate HTML Summary for EnrichGT Object}
\usage{
egt_summary(x, name)
}
\arguments{
\item{x}{An EnrichGT_obj object}

\item{name}{The cluster name to summarize}
}
\value{
HTML formatted summary that can be viewed with htmltools::html_print() or similar
}
\description{
This function generates an HTML-formatted summary of enrichment results
that can be displayed in RStudio viewer or web browsers.
}
\examples{
\dontrun{
# Generate HTML summary for cluster 1
html_summary <- egt_summary(obj, "Cluster_1")
htmltools::html_print(html_summary)

# Or view in RStudio viewer
if (interactive()) {
  html_summary <- egt_summary(obj, "Cluster_1")
  print(html_summary)
}
}
}
