% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{egtPlot}
\alias{egtPlot}
\title{Visualize results generated form `EnrichGT()` using ggplot2}
\usage{
egtPlot(x, ...)
}
\arguments{
\item{x}{an EnrichGT object}

\item{...}{}

\item{ntop}{Show top N in each cluster}

\item{low.col}{the color for the lowest}

\item{hi.col}{the color for the highest}
}
\value{
a ggplot2 object
}
\description{
Visualize results generated form `EnrichGT()` using ggplot2
}
\author{
Zhiming Ye
}
