% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{EnrichGT}
\alias{EnrichGT}
\title{Parse enrichment results and further clustering and visualizing}
\usage{
EnrichGT(x, ClusterNum = 15, P.adj = 0.05, force = F, nTop = 10, ...)
}
\arguments{
\item{x}{an enrichment result from `clusterProfiler`, or a `data.frame` containing result from `clusterProfier`}

\item{ClusterNum}{how many cluster will be clustered}

\item{P.adj}{p.adjust cut-off. To avoid slow visualization, you can make stricter p-cut off.}

\item{force}{ignore all auto-self-checks, which is useful}

\item{nTop}{keep n top items according to p-adj in each cluster.}

\item{...}{Others options.}
}
\value{
an `EnrichGT_obj` object. slot `enriched_result` contains a data.frame with enriched results. `gt_object` contains `gt` object. you can use `obj@gt_object` to get it and use functions from `gt` like `gtsave`. `gene_modules` is a list containing meta-gene modules of each cluster. `pathway_clusters` contains pathways names in each cluster.
}
\description{
Parse enrichment results and further clustering and visualizing
}
\details{
For an ORA result, c("ID","Description","GeneRatio","pvalue","p.adjust","geneID","Count") should be contained; For GSEA, c("ID","Description","NES","pvalue","p.adjust","core_enrichment") should be contain. For `compareClusterResult`, a `compareClusterResult` object or a data-frame with additional `Cluster` column should be contained, others similar to ORA result.
}
\author{
Zhiming Ye
}
