% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_helpers.R
\name{genes_with_weights}
\alias{genes_with_weights}
\title{Return ranked gene list which is use for "GSEA" or other places}
\usage{
genes_with_weights(genes, weights)
}
\arguments{
\item{genes}{A vector containing genes}

\item{weights}{A vector contain weight of genes, typically like log2FC from DEG analysis}
}
\value{
A ranked named numeric vector. Names of the numbers is the ENTREZID.
}
\description{
Return ranked gene list which is use for "GSEA" or other places
}
\author{
Zhiming Ye
}
