% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{database_GO_BP}
\alias{database_GO_BP}
\alias{database_GO_CC}
\alias{database_GO_MF}
\alias{database_GO_ALL}
\alias{database_Reactome}
\alias{database_progeny_human}
\alias{database_progeny_mouse}
\alias{database_CollecTRI_human}
\alias{database_CollecTRI_mouse}
\title{Get database form Gene ontology or Reactome Pathways}
\usage{
database_GO_BP(OrgDB = org.Hs.eg.db)

database_GO_CC(OrgDB = org.Hs.eg.db)

database_GO_MF(OrgDB = org.Hs.eg.db)

database_GO_ALL(OrgDB = org.Hs.eg.db)

database_Reactome(OrgDB = org.Hs.eg.db)

database_progeny_human()

database_progeny_mouse()

database_CollecTRI_human()

database_CollecTRI_mouse()
}
\arguments{
\item{OrgDB}{org.DB form bioconductor, can be org.Hs.eg.db or org.Mm.eg.db,... GO and Reactome should add this, progeny and collectri do not.}
}
\value{
a data.frame with ID, terms and genes
}
\description{
Get database form Gene ontology or Reactome Pathways
}
