% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_helpers.R
\name{getGMTFile}
\alias{getGMTFile}
\title{getGMTFile}
\usage{
getGMTFile(gmtfile)
}
\arguments{
\item{gmtfile}{gmt file}
}
\value{
data.frame
}
\description{
read `.gmt` files
}
\details{
parse gmt file to a data.frame
}
\author{
cited from https://github.com/YuLab-SMU/gson/blob/main/R/GMT.R
}
