% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_helpers.R
\name{Ranked.GS}
\alias{Ranked.GS}
\title{Return ranked gene list which is use for "GSEA"}
\usage{
Ranked.GS(Gene, Weight)
}
\arguments{
\item{Gene}{A vector containing genes}

\item{Weight}{A vector contain weight of genes, typically like log2FC from DEG analysis}
}
\value{
A ranked named numeric vector. Names of the numbers is the ENTREZID.
}
\description{
Return ranked gene list which is use for "GSEA"
}
\author{
Zhiming Ye
}
