% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_helpers.R
\name{database_from_gmt}
\alias{database_from_gmt}
\title{Parse GMT format gene set files}
\usage{
database_from_gmt(gmtfile, OrgDB = NULL, convert_2_symbols = T)
}
\arguments{
\item{gmtfile}{Path to GMT format file}

\item{OrgDB}{Annotation database for ID conversion (e.g., org.Hs.eg.db for human).
Required if convert_2_symbols=TRUE.}

\item{convert_2_symbols}{Logical indicating whether to convert ENTREZ IDs to gene symbols.
Default is TRUE.}
}
\value{
A data frame with columns:
  \itemize{
    \item term: Gene set name
    \item gene: Gene identifiers (symbols or ENTREZ IDs)
  }
  If input has 3 columns, includes an additional ID column.
}
\description{
Reads gene set files in GMT format (e.g., from MSigDB or WikiPathways) and converts
them to a data frame suitable for enrichment analysis. Can optionally convert ENTREZ IDs
to gene symbols.
}
\examples{
\dontrun{
# Read MSigDB hallmark gene sets
gmt_file <- system.file("extdata", "h.all.v7.4.symbols.gmt", package = "EnrichGT")
gene_sets <- database_from_gmt(gmt_file)

# Read WikiPathways with ENTREZ to symbol conversion
gmt_file <- "wikipathways-20220310-gmt-Homo_sapiens.gmt"
gene_sets <- database_from_gmt(gmt_file, OrgDB = org.Hs.eg.db)
}
}
\author{
Original GMT parser by Guangchuang Yu (https://github.com/YuLab-SMU/gson).
  Cache system and enhancements by Zhiming Ye.
}
