% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2.R
\name{egt_fetch_termwise_relationship}
\alias{egt_fetch_termwise_relationship}
\title{Fetch and visualize termwise relationships in enrichment analysis results}
\usage{
egt_fetch_termwise_relationship(
  x,
  database,
  according_to = "Description",
  ClusterNum = 4,
  method = "ward.D2",
  force = F,
  maxLength = 40
)
}
\arguments{
\item{x}{A character vector of terms to analyze}

\item{database}{Query database. For example, `database_GO_BP(org.Hs.eg.db)`. If you are quering fused result, you can input a list like `list(database_GO_BP(org.Hs.eg.db), database_Reactome(org.Hs.eg.db))`.}

\item{according_to}{Character string specifying which column to use for filtering, default is "Description"}

\item{ClusterNum}{Integer specifying the number of clusters, default is 4}

\item{method}{Character string specifying the hierarchical clustering method, default is "ward.D2"}

\item{force}{Avoid checking the length of x. Large query will cause long calculation time and poor results.}

\item{maxLength}{Warp length max than what terms}
}
\value{
A list containing:
  \item{Figure}{A ggplot2 object showing the hierarchical clustering}
  \item{Cluster_DF}{A data frame with cluster assignments}
  \item{EnrichedHC}{The hierarchical clustering results}
}
\description{
This function analyzes the relationships between enriched terms and creates a
hierarchical clustering visualization. Please avoiding querying too large dataset.
}
\examples{
\dontrun{
# Example usage:
result <- egt_fetch_termwise_relationship(
  x = c("pathway1", "pathway2"),
  database = pathway_db,
  according_to = "Description"
)
result$Figure
}

}
