% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{EnrichGT}
\alias{EnrichGT}
\title{Parse enrichment results and further clustering and visualizing}
\usage{
EnrichGT(x, ClusterNum = 15, P.adj = 0.05, ...)
}
\arguments{
\item{x}{an enrichment result from `clusterProfiler`, or a `data.frame` containing result from `clusterProfier`}

\item{ClusterNum}{how many cluster will be clustered}

\item{P.adj}{p.adjust cut-off. To avoid slow visualization, you can make stricter p-cut off.}

\item{...}{Others options.
For an ORA result, c("ID","Description","GeneRatio","pvalue","p.adjust","geneID","Count") should be contained; For GSEA, c("ID","Description","NES","pvalue","p.adjust","core_enrichment") should be contain. For `compareClusterResult`, a `compareClusterResult` object or a data-frame with additional `Cluster` column should be contained, others similar to ORA result.}
}
\value{
a `gt` great table object or a `List` containing multiple `gt` objects for `compareClusterResult`.
}
\description{
Parse enrichment results and further clustering and visualizing
}
\author{
Zhiming Ye
}
