% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{database_...}
\alias{database_...}
\alias{database_GO_BP}
\alias{database_GO_CC}
\alias{database_GO_MF}
\alias{database_GO_ALL}
\alias{database_Reactome}
\alias{database_progeny_human}
\alias{database_progeny_mouse}
\alias{database_CollecTRI_human}
\alias{database_CollecTRI_mouse}
\title{Get database for enrichment analysis}
\usage{
database_GO_BP(OrgDB = org.Hs.eg.db)

database_GO_CC(OrgDB = org.Hs.eg.db)

database_GO_MF(OrgDB = org.Hs.eg.db)

database_GO_ALL(OrgDB = org.Hs.eg.db)

database_Reactome(OrgDB = org.Hs.eg.db)

database_progeny_human()

database_progeny_mouse()

database_CollecTRI_human()

database_CollecTRI_mouse()
}
\arguments{
\item{OrgDB}{The AnnotationDbi database to fetch pathway data and convert gene IDs to gene symbols. For human it would be `org.Hs.eg.db`, for mouse it would be `org.Mm.eg.db`. In AnnotationDbi there are many species, please search `AnnotationDbi` for other species annotation database. GO and Reactome should add this, progeny and collectri do not.}
}
\value{
a data.frame with ID, terms and genes
}
\description{
Get Gene Ontology (GO), Reactome, and other term-to-gene database, for enrichment analysis
}
\author{
Zhiming Ye. Part of functions were inspired by `clusterProfiler` but with brand new implement.
}
